Optimise brewing stands
Cache the isBrewable result as the operation to check mixes can be quite heavy. Especially when being ticked on every single tick. The cache gets invalidated when the block entity is changed. Fortunately inventory updates trigger the block entity's change function.

This has a big improvement for servers that have a lot of active brewing stands.