Optimise ticking block entities
This one is more of a theory. In my tests I did see a 3mspt decrease, but it was with 90k block entities across 121 chunks.

I did have another test where the result was ~0.5mspt decrease. So it really depends on how many block entities you have & if players have dense bases.