Cached checks for special dates
Based off the one that Pufferfish had for Bats. Though 1.21.11 removed that whole function from Bats. I decided to add it back as zombies and skeletons still do Halloween checks. Though this time adding support for Christmas for if Mojang decides to make the server do something special on Christmas in a future version.